+++ sources/packages/backend/apps/api/src/route-controllers/student/student.aest.controller.ts
@@ -72,6 +72,7 @@ import {
   SIN_VALIDATION_RECORD_NOT_FOUND,
 } from "../../constants";
 import { Role } from "../../auth/roles.enum";
+import { EntityManager } from "typeorm";
 
 /**
  * Student controller for AEST Client.
@@ -198,15 +199,16 @@ export class StudentAESTController extends BaseController {
 
     // This method will be executed alongside with the transaction during the
     // execution of the method updateStudentFiles.
-    const sendFileUploadNotification = () =>
+    const sendFileUploadNotification = (entityManager: EntityManager) =>
       this.notificationActionsService.sendMinistryFileUploadNotification(
         {
           firstName: student.user.firstName,
           lastName: student.user.lastName,
           toAddress: student.user.email,
+          userId: student.user.id,
         },
-        student.user.id,
         userToken.userId,
+        entityManager,
       );
     // Updates the previously temporary uploaded files.
     await this.fileService.updateStudentFiles(
