+++ sources/packages/backend/apps/api/src/services/student-file/student-file.service.ts
@@ -123,9 +123,7 @@ export class StudentFileService extends RecordDataModelService<StudentFile> {
     uniqueFileNames: string[],
     fileOrigin: FileOriginType,
     groupName: string,
-    sendNotification?: (
-      entityManager: EntityManager,
-    ) => Promise<GCNotifyResult>,
+    sendNotification?: (entityManager: EntityManager) => Promise<void>,
     metadata?: StudentFileMetadata,
   ): Promise<UpdateResult> {
     let updateResult: UpdateResult;
