+++ sources/packages/backend/libs/services/src/notifications/notification/notification.model.ts
@@ -8,3 +8,26 @@ export interface SaveNotificationModel {
     template_id: string;
   };
 }
+
+export interface StudentFileUploadNotification {
+  firstName: string;
+  lastName: string;
+  birthDate: Date;
+  applicationNumber?: string;
+  documentPurpose: string;
+  userId: number;
+}
+
+export interface MinistryStudentFileUploadNotification {
+  firstName: string;
+  lastName: string;
+  toAddress: string;
+  userId: number;
+}
+
+export interface StudentRestrictionAddedNotification {
+  givenNames: string;
+  lastName: string;
+  toAddress: string;
+  userId: number;
+}
